Move CompositeSignaturesWithKeys into net.corda.core.crypto package.
(cherry picked from commit 8f29562)
